backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9
pydantic>=2.7
python-dotenv>=1.0
google-generativeai>=0.5.0
camelot-py[cv]>=0.11.0
pandas>=2.0.0